requirements.txt
pillow
tensorflow-cpu
Flask
gunicorn
